Fix to support returning all fields and numbers
If the first record found has fields that are null, socrata doesn’t
return the field, so it is missing in the feature service json.  This
fix ensures all fields are there.  It also return numeric fields as
numbers instead of strings.  We are also storing the fields and socrata
field types for better date handling in the future.